@@ -44,5 +44,5 @@ subprojects {
     }
 
     sourceCompatibility = 1.8
-    targetCompatibility = 1.8
+    targetCompatibility = 1.11
 }
